requirement.txt
Django
django-multiselectfield
oauthlib
Pillow
PyJWT
PyMySQL
PyMysqlDB
python-alipay-sdk
python3-openid
pytz
rauth
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
urllib3
django-haystack
whoosh
jieba
djangorestframework
djangorestframework-jwt
drf-dynamic-fields
awesome-slugify
coreapi
django-cors-headers
django-pdfkit
cffi
django-rest-knox
django-filter